WIP: Rename BaseStage to StageBase
