+++ morpheus/pipeline/stage_base.py
